test: validating failure
